xbee/ZigBeeBell/requirements.txt
XBee
pyserial
